# Improve master_commit_red query performance